allow Collector to count duplicated cards
There's an ambiguity in the way the Collector interacts with wild cards.
The bonus text says "different cards", but it's not immediately clear
what that means. I.e., if you have one card in a suit and then duplicate
it with a Doppelgänger, Shapshifter, or Mirage, have you now collected
two items in that suit, or still only one?

The argument for two items is:

  - you now have two physically different cards which both have that
    suit

  - that's how the WizKids app scores it

The argument for one item is:

  - the cards are not "different" cards, in that they have the same name

  - Bruce Glassco says so (multiple times):

      - https://boardgamegeek.com/thread/1838508/article/26773487#26773487

      - https://boardgamegeek.com/thread/1911765/article/28414416#28414416

I wrote this patch after researching how the WizKids app scores it. But
now after seeing Bruce's comments, I'm inclined to say it is the wrong
thing (and your scoring is correct and WizKids has a bug). However, I
thought I'd throw it out here for posterity and the sake of discussion.
Possibly it could become a "house rule" option, but I'm not sure you'd
want to clutter up the interface with that.